Revert "Adityas branch"
